dhcp/dhclient: emit a warning if the DHCPv6 hostname is not a FQDN
RFC 4704 ("The DHCPv6 Client FQDN Option") allows both partial and
fully-qualified names in the FQDN option, however dhclient always
appends a terminating zero-length label to the name, so we ignore
unqualified hostnames to prevent a wrong configuration.

Emit a warning when the field is ignored so that users can clearly see
why the hostname is not being sent.