version: remove NM_AVAILABLE_IN_1_4 macros and add NM_AVAILABLE_IN_1_2_4
The NM_AVAILABLE_IN_* and NM_DEPRECATED_IN_* macros are kinda internal macros
to mark public API (although they are in public headers).

On 1.2 branch, we will never add a new API that is marked with one of the
1.4 macros, because from 1.2's point of view 1.4 does not exist. Remove
those macros, but add instead a 1.2.4 macro that we will use.