ethernet: cleanup type definition of NMDeviceEthernet
No longer typedef NMDeviceEthernet to NMDevice. We don't do that
for most other classes, and I think it is not a good pattern
(yes, the casts are cumbersome, but what can you do).

Also, embed a pointer to the private data in NMDeviceEthernet
for fast lookup and ease of debugging.