policy: fix hostname change when IP method ends after activation (take 2)
Commit 9446481f4cc0 ("policy: update system hostname when DHCP
configuration changes") tried to fix the missing hostname change when
IPv4 receives a hostname through DHCP but terminates after IPv6, by
calling update_routing_and_dns() as soon as the new DHCP configuration
was received.

It turns out that doing so is not always effective because the device
must be the "best" device (the one with default route) in order to
trigger a hostname change, but the best device status is decided
later. Updating the hostname in device_ipx_config_changed() should
cover all cases.

Fixes: 9446481f4cc06c657d9467bce0921a85bb94ae6b
https://bugzilla.redhat.com/show_bug.cgi?id=1356015
https://bugzilla.redhat.com/show_bug.cgi?id=1364393