core: don't suppress routes inside the subnet of the interface
It's not clear why a route should be suppressed if it is contained
in the subnet of one of the interface's addresses.

I think it is wrong to do this. For example, imagine an ethernet
and a Wi-Fi device both connected to the same subnet 10.0.0.0/8. By
default, ethernet gets higher priority and a better metric of 100.
If the user wants to configure a route "10.0.0.1/32 metric 99"
to reach a certain host explicitly via Wi-Fi, this check will
forbid that.

This condition was added a long time ago (38dbdae26606f984f1d12c8a56ed27bc1b4d9c20),
but it's unclear what the original intent was.

See also commit 4f7b1cabc063bfda96cb5c129d6a233e5d5cff68, which
already relaxed this suppression of routes for non-direct routes.

(cherry picked from commit ac5dc1a9510c086c54c365b1160a51cc25402010)